pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = "AI 에이전트 프로젝트"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
openai = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
streamlit = "^1.31.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 